Move Gatherers plugin management to gatherers package
